Simplify user selection logic - always use email to select target user.
So it turns out that, even with username-based federated IDs, you can always select a user to act on by specifying the user's email.  This allows both forward and reverse directions to use the email as the user specifier, which simplifies the logic (and probably makes it more robust).

Also added these enhancements:

* Make sure that the input row is valid before trying to execute it.
* Add test data based on the actual export format for users from the dashboard, to make sure the additional columns don't interfere with the CSV extract logic.